refactor dir tree & add event bus
