+++ src/main/java/com/example/SBE/Blade.java
@@ -1,4 +1,4 @@
-package com.example.examplemod;
+package com.example.SBE;
 
 import java.util.List;
