+++ src/main/java/com/example/SBE/BladeRecipe.java
@@ -1,4 +1,4 @@
-package com.example.examplemod;
+package com.example.SBE;
 
 import mods.flammpfeil.slashblade.ItemSlashBladeNamed;
 import mods.flammpfeil.slashblade.SlashBlade;
