+++ src/main/java/com/example/SBE/RecipePos.java
@@ -1,4 +1,4 @@
-package com.example.examplemod;
+package com.example.SBE;
 
 public class RecipePos {
     private int posX = 0;
