+++ src/main/java/com/example/SBE/SBE.java
@@ -1,6 +1,6 @@
-package com.example.examplemod;
+package com.example.SBE;
 
-import net.minecraft.init.Blocks;
+import mods.flammpfeil.slashblade.SlashBlade;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
@@ -8,8 +8,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import org.apache.logging.log4j.Logger;
 
 @Mod(modid = SBE.MODID, name = SBE.NAME, version = SBE.VERSION)
-public class SBE
-{
+public class SBE {
     public static final String MODID = "slashBladeEx";
     public static final String NAME = "SlashBladeExtension";
     public static final String VERSION = "1.0";
@@ -18,7 +17,7 @@ public class SBE
 
     @EventHandler
     public void preInit(FMLPreInitializationEvent event) {
-
+        SlashBlade.InitEventBus.register(new BladeRegister());
     }
 
     @EventHandler
