+++ src/main/java/com/example/SBE/blade/nihil.java
@@ -1,4 +1,4 @@
-package com.example.examplemod.blade;
+package com.example.SBE.blade;
 
 import mods.flammpfeil.slashblade.ItemSlashBladeNamed;
 import mods.flammpfeil.slashblade.RecipeAwakeBlade;
