+++ src/main/resources/assets/nihil/SBE/model/nihil/nihil.obj
